chore: improve general performance
