+++ src/bin/index.ts
@@ -1,7 +1,6 @@
 #! /usr/bin/env node
 
 import type { Configs } from '../@types/poku.js';
-import { argv, exit } from 'node:process';
 import { escapeRegExp } from '../modules/helpers/list-files.js';
 import { getArg, getPaths, hasArg, argToArray } from '../parsers/get-arg.js';
 import { states } from '../configs/files.js';
@@ -18,7 +17,7 @@ import { getConfigs } from '../parsers/options.js';
     const { VERSION } = require('../configs/poku.js');
 
     Write.log(VERSION);
-    exit(0);
+    return;
   }
 
   const enforce = hasArg('enforce') || hasArg('x', '-');
@@ -108,61 +107,13 @@ import { getConfigs } from '../parsers/options.js';
     Write.log(`Total test files: ${format(String(files.length)).bold()}`);
     Write.hr();
 
-    exit(0);
+    return;
   }
 
   if (enforce) {
-    const allowedFlags = new Set([
-      '--config',
-      '--platform',
-      '--filter',
-      '--exclude',
-      '--killPort',
-      '--killRange',
-      '--killPid',
-      '--denoAllow',
-      '--denoDeny',
-      '--denoCjs',
-      '--parallel',
-      '--enforce',
-      '--quiet',
-      '--debug',
-      '--failFast',
-      '--watch',
-      '--envFile',
-      '--concurrency',
-      '--watchInterval',
-      '--node',
-      '--bun',
-      '--deno',
-      '-c',
-      '-p',
-      '-d',
-      '-q',
-      '-w',
-      '-x',
-    ]);
+    const { checkFlags } = require('./enforce.js');
 
-    const args = argv.slice(2);
-    const unrecognizedFlags: string[] = [];
-
-    for (const arg of args) {
-      const flagName = arg.split('=')[0];
-
-      if (!allowedFlags.has(flagName) && flagName.startsWith('-')) {
-        unrecognizedFlags.push(flagName);
-      }
-    }
-
-    if (unrecognizedFlags.length > 0) {
-      Write.hr();
-      Write.log(
-        `${format('Unrecognized flags:').bold()}\n\n${unrecognizedFlags.map((flag) => format(flag).fail()).join('\n')}`
-      );
-      Write.hr();
-
-      exit(1);
-    }
+    checkFlags();
   }
 
   const tasks: Promise<unknown>[] = [];
